Move ExponentialWaitStrategy to SPI so core:base does not need to be referenced
